moved callbacks away from pages
trying out splitting the callbacks for a page into a neighboring folder to separate layout logic (pages) from data logic (callbacks)
